Refactored def info
