+++ external-crates/move/crates/move-analyzer/src/completions/utils.rs
@@ -6,13 +6,14 @@ use std::path::PathBuf;
 use crate::symbols::{
     compute_symbols_parsed_program, compute_symbols_pre_process,
     cursor::CursorContext,
+    def_info::DefInfo,
     ide_strings::{
         mod_ident_to_ide_string, ret_type_to_ide_str, type_args_to_ide_string,
         type_list_to_ide_string,
     },
     types::{
-        AutoImportInsertionInfo, AutoImportInsertionKind, CompiledPkgInfo, DefInfo, ModuleDefs,
-        Symbols, SymbolsComputationData,
+        AutoImportInsertionInfo, AutoImportInsertionKind, CompiledPkgInfo, ModuleDefs, Symbols,
+        SymbolsComputationData,
     },
 };
 use lsp_types::{
