+++ external-crates/move/crates/move-analyzer/src/symbols/ide_strings.rs
@@ -4,7 +4,7 @@
 //! This module contains the implementation of functions supporting conversion of various
 //! constructs to their IDE-friendly string representations.
 
-use crate::symbols::types::{FunType, VariantInfo};
+use crate::symbols::def_info::{FunType, VariantInfo};
 use move_compiler::{
     expansion::{
         ast::{self as E, AbilitySet, ModuleIdent_, Value, Value_, Visibility},
